Backend/requirements.txt
Django==3.2
mysqlclient==2.0.3
djangorestframework==3.12.4
django-cors-headers==3.7.0  
